Fix lockup on JVM exit caused by push over SSH
Push does not clean up the Transport correctly when it is done
so the JSch library still has user level threads running and we
do not close down the JVM.  We must close the transport.

This was broken by 2e05675201 ("Reuse the same SSH connection")
as the JSch session is held open inside of the Transport instance
in case there are additional calls made on the same Transport.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>