Add parseTree method to RevWalk to obtain a RevTree from AnyObjectId
Some applications want to dereference an ObjectId back to a tree,
and they don't care what type of object the ObjectId supplied is
actually of.  If its an annotated tag we can peel the tag back to
either a tree, or a commit, and a commit can be peeled back to its
only tree.  So parseTree is a simple API for ^{tree}.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>