Add option handler for AbstractTreeIterator values
Converts String to AbstractTreeIterator on demand.  A string
that can be resolved as a RevTree is read with a canonical
tree iterator, as it is a standard Git tree already stored in
the object database.  Otherwise we assume it is a local file
and scan the filesystem, as it may be a working directory.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>