Switch jgit.pgm to J2SE-1.5 execution environment
We have been keeping the jgit library itself on Java 5, and the
jgit command line tools should also be on Java 5 and avoid using
any Java 6 APIs (for now).  Not all of our target platforms have
a Java 6 virtual machine available out of the box.

Since the pgm project broke out of the library project our code
already conforms to Java 5 APIs, so we just have to switch the
build path settings.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>