Create a lightweight registration wrapper for TextBuiltin
Automatic command list generation requires knowing what commands
are available to this runtime, and what name those commands can be
called as by the end-user.  This lightweight wrappers carries the
data from the Command annotation, possibly filling in the name of
the command by generating it from the implementation class name.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>