Rename INSTALL to EGIT_INSTALL
This document is entirely about how to install the EGit plugins
and has nothing to do with JGit itself.  Later if/when we split the
Git repository into two trees (one for jgit, one for egit) we can
rename this back to just plain INSTALL inside of the EGit repository.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>