Remove the "windows isn't supported" remark from README
The Windows support has come a long way since this was written
into the README.  We actually (mostly) support repositories that
were created on Cygwin, even if they used --shared to setup object
alternates with Cygwin paths, and we honor the user's Cygwin home
directory for configuration and SSH keys.  Windows is about as well
supported as one can get on Java.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>