Fix RepositoryMapping.getRepoRelativePath to honor linked resources
If a project resource is linked to a repository working directory we
should be locating the path of the resource within the repository by
the path it appears in the filesystem, not the path it appears at in
the Eclipse project structure.  By using the filesystem location we
can ensure the paths match what C Git would see when accessing the
same resource.

By making this change we can now safely remove the subset concept
from RepositoryMapping as it was a crude form of trying to come up
with the same result when projects appeared within a repository.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>