Allow WorkingTreeIterators to define their prefix path when created
When iterating a subset of the working directory over the full
repository tree (e.g. a CanonicalTreeParser run in parallel in
the same TreeWalk) we need the working directory to know what
prefix it must apply to all of its own generated paths so they
match up with the paths of the CanonicalTreeParser coming from
the object database.

The prefix is only set on the root level, as we only need to
inject it into the first iterator.  After that the shared path
buffer will ensure the subtree iterators (if any are created)
will have the proper path in them too.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>