Rely upon Constants.CHARSET over Constants.CHARACTER_ENCODING
This avoids the UnsupportedEncodingException error which we know
must not fail on a JRE that supports jgit.  We have to have the
stock UTF-8 encoding available for basic operation.

In some cases we can also avoid an unnecessary copy.  Some code
paths in the byte[]->String utility methods in the Sun JRE seem
to copy the input byte[] prior to calling the character set for
the conversion.  This copy is unnecessary.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>