Allow AbstractTreeIterators to find out about StopWalkExceptions
If a tree iterator has overridden skip() to specially handle entries
which the TreeWalk has decided are not relevant then it might also
need to pay attention to a StopWalkException.  The SWE is thrown out
of a filter and behaves much like skip, except it skips everything
else in the tree.  This is used by PathFilterGroup to break out of
a walk once we know that no additional records could ever match.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>