Support locking (and unlocking) a .git/index through DirCache
In order to safely perform an update of the .git/index file (or any
other DirCache file) we need to lock the file prior to reading its
contents into memory.  This ensures that the contents will not be
modified between the time we start processing the entires and when
the update operation is completed.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>