Support recursively getting all entries under a subtree path
Some applications need to process through all entries of a subtree.
A good example is if the entries need to be recreated into another
subtree to support a "cp -r a b" or "mv a b" operation.

The getEntriesWithin method returns an array of all entries within
a single subtree path, allowing the aplication to process over that
array in order.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>