Add a basic command line implementation of rm
Removing entries from the index and from the working directory is a
very trivial operation now that we have DirCacheBuilder able to do
an update of the index, automatically keeping the paths that were
skipped over by the TreeWalk.  Therefore we only need to match the
paths the user wants to remove, and do nothing with them.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>