Rewrite UntrackOperation to use DirCacheBuilder
We can more efficiently untrack single files and entire directories using
the DirCacheBuilder API as that API is able to perform range edits on
the index data structure.  Its also able to perform faster disk IO, due
to less conversion overheads.

Although performance of untrack is not a critical feature it is a good
idea to move it over to the newer API, as we will eventually get rid of
the GitIndex class entirely.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>