Allow EGit to handle linked resources at any depth
We already were handling linked resources that appear anywhere
in the project, but Eclipse didn't know we were willing to do
that sort of magic in our provider code as we did not override
the correct method for it in GitProvider.

We now support linked folders which are at any level.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>