Add a TreeWalk iterator implementation for IContainer
Treating the Eclipse workspace as a filesystem backend for TreeWalk
can give us some nice caching on the file modification times and on
directory contents.  We also get nice APIs to open a file and read
its contents.

This iterator allows combining a walk over an IContainer with any
other sort of directory walk we may do, like against a stored tree
object in the object database or against the index file, or even
any other IContainer in the workspace (e.g. diff two projects).

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>

Dropped the use of IContainer.INCLUDE_HIDDEN.

Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>