Paper bag fix 'jgit glog' handling of commit-ish arguments
When we parsed these arguments into RevCommit instances they were
done against the wrong RevWalk instance.  We parsed them into a
generic RevWalk which has no plotting support, so the objects do
not have the extra fields used by the plotting code.  We need to
lookup the commit from the PlotRevWalk and use those instances.

One of the calls (markStart) failed to use the secondary RevWalk
instance, resulting in a ClassCastException during plotting.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>