Add simple abbreviate() method to ObjectId
It's common usage to search for short, but unique abbreviation of
object's SHA-1, so let's have a method for that.

This is a really workaround implementation, but it has been already used
in other places. Let's have other usages pointing here, so we can easily
change only this implementation later.

Signed-off-by: Marek Zawirski <marek.zawirski@gmail.com>

[Amended not to break Branch command in latest master]
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>