Set FileDialog selection appropriately in clone wizard
Previous selection scheme was probably not correct, or at least not
portable. It hasn't worked on GTK+ SWT implementation.

New code still doesn't work on GTK+ SWT in Eclipse 3.3, but it works in
Eclipse 3.4. FileDialog for GTK+ in 3.3 was probably buggy. I've created
a bug at eclipse.org bugzilla for that (241824), but it won't be fixed
for 3.3.

Signed-off-by: Marek Zawirski <marek.zawirski@gmail.com>
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>