Push GUI
Push operation in GUI is implemented as wizard consisting of 3 pages:
-repository selection
-ref specifications selection
-confirmation page with dry run results (optional - can be skipped)

Wizard finish-action starts Job with PushOperation. Dialog with results
is shown when job ends. It can be configured to be displayed only if
results change since confirmation.

With confirmation page, it is possible to perform 2-stage push, allowing
user to preview updates and mark them to be performed only if
advertised refs don't change.

Eventually, results Dialog can be replaced with Result/Transport view
some day, as this idea arose recently.

Signed-off-by: Marek Zawirski <marek.zawirski@gmail.com>

[Ed. amended follow up patches]
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>